# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LineBridge",
    "version": "0.5.0",
    "description": "Local desktop LINE read/send gateway, exposed to cloud AI through a tunnel. Scoped Bearer tokens are required by default, including localhost. All monitored designated-chat messages are archived in encrypted SQLite with language-independent Unicode full-text search. Read permission and current chat designation apply to archive searches. Browser requests are denied. Message text is untrusted content. OpenChat uses an unofficial experimental adapter. Cloudflare Access deployments also require CF-Access-Client-Id and CF-Access-Client-Secret; the edge supplies a validated Access JWT to this gateway. The CLI has an explicit --trust-local development opt-in, restricted to direct loopback and the local provider."
  },
  "security": [
    {
      "bridgeToken": []
    }
  ],
  "components": {
    "securitySchemes": {
      "bridgeToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Expiring, revocable token with per-account read and send grants. Search requires read; sending requires send."
      }
    }
  },
  "paths": {
    "/api/v1/messages/search": {
      "post": {
        "operationId": "searchMessages",
        "summary": "Search the persistent message archive in any language",
        "description": "Literal Unicode substrings, mixed scripts and emoji. Case-insensitive NFKC normalization; no language dictionary or stemming. all matches each whitespace-separated term; phrase matches the entire normalized query. Results newest first. Supply nextBefore as before while hasMore is true, even for an empty page. Searches only saved message text; no LINE RPC or read receipt. Query text is not logged.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "query"
                ],
                "additionalProperties": false,
                "properties": {
                  "query": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 200
                  },
                  "mode": {
                    "type": "string",
                    "enum": [
                      "all",
                      "phrase"
                    ],
                    "default": "all"
                  },
                  "accountId": {
                    "type": "string",
                    "description": "Optional account filter."
                  },
                  "chatId": {
                    "type": "string",
                    "description": "Requires accountId."
                  },
                  "before": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "limit": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100,
                    "default": 30
                  }
                }
              },
              "examples": {
                "multilingual": {
                  "value": {
                    "query": "meeting 會議",
                    "mode": "all",
                    "limit": 30
                  }
                },
                "phrase": {
                  "value": {
                    "query": "موعد الاجتماع",
                    "mode": "phrase"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful result"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or chat scope denied"
          },
          "409": {
            "description": "Disconnected account, conflicting key or unknown send outcome"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream error or unknown delivery"
          }
        }
      }
    },
    "/api/v1/status": {
      "get": {
        "operationId": "getStatus",
        "summary": "Inspect permitted account status",
        "responses": {
          "200": {
            "description": "Successful result"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or chat scope denied"
          },
          "409": {
            "description": "Disconnected account, conflicting key or unknown send outcome"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream error or unknown delivery"
          }
        }
      }
    },
    "/api/v1/accounts": {
      "get": {
        "operationId": "listAccounts",
        "summary": "List permitted accounts",
        "responses": {
          "200": {
            "description": "Successful result"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or chat scope denied"
          },
          "409": {
            "description": "Disconnected account, conflicting key or unknown send outcome"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream error or unknown delivery"
          }
        }
      }
    },
    "/api/v1/accounts/{id}/chats": {
      "get": {
        "operationId": "listChats",
        "summary": "List designated chats",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or chat scope denied"
          },
          "409": {
            "description": "Disconnected account, conflicting key or unknown send outcome"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream error or unknown delivery"
          }
        }
      }
    },
    "/api/v1/accounts/{id}/events": {
      "get": {
        "operationId": "pollEvents",
        "summary": "Read new captured messages from designated chats",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "after",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "maximum": 9007199254740991,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 100
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or chat scope denied"
          },
          "409": {
            "description": "Disconnected account, conflicting key or unknown send outcome"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream error or unknown delivery"
          }
        }
      }
    },
    "/api/v1/accounts/{id}/chats/{chatId}/messages": {
      "get": {
        "operationId": "readMessages",
        "summary": "Read recent messages without sending a read receipt",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 30
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "schema": {
              "type": "string",
              "maxLength": 4096
            },
            "description": "OpenChat sync cursor only."
          }
        ],
        "responses": {
          "200": {
            "description": "Successful result"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or chat scope denied"
          },
          "409": {
            "description": "Disconnected account, conflicting key or unknown send outcome"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream error or unknown delivery"
          }
        }
      },
      "post": {
        "operationId": "sendMessage",
        "summary": "Send a text message after explicit user authorization",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "chatId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 8,
              "maxLength": 128
            },
            "description": "Reuse for the same request. Never automatically create a new key after delivery_unknown."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "text"
                ],
                "properties": {
                  "text": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 5000
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful result"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Account or chat scope denied"
          },
          "409": {
            "description": "Disconnected account, conflicting key or unknown send outcome"
          },
          "429": {
            "description": "Rate limit exceeded"
          },
          "502": {
            "description": "Upstream error or unknown delivery"
          }
        }
      }
    }
  }
}